clay: introduce special conversion path %mime -> b in +build-cast
External edits in clay accept %mime data without requiring
the %mime mark to be present in the desk.
To introduce this feature for internal edits, a new conversion
path is added to +build-cast, so that if a desk is missing
the %mime mark, a file can still be converted using
the +grow arm of the target mark.